+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For a full diff see [`2.28.3...main`][2.29.0...main].
 
 - Started injecting `Printer\Printer` instead of `Formatter\Formatter` into `NormalizeCommand` ([#1008]), by [@ergebnis-bot]
 - Required `composer/composer:2.5.1` for compiling `composer-normalize.phar` ([#1020]), by [@localheinz]
+- Required `ergebnis/json-normalizer:^4.0.0` ([#1056]), by [@dependabot]
 
 ## [`2.29.0`][2.29.0]
 
@@ -1000,6 +1001,7 @@ For a full diff see [`81bc3a8...0.1.0`][81bc3a8...0.1.0].
 [#1004]: https://github.com/ergebnis/composer-normalize/pull/1004
 [#1008]: https://github.com/ergebnis/composer-normalize/pull/1008
 [#1020]: https://github.com/ergebnis/composer-normalize/pull/1020
+[#1056]: https://github.com/ergebnis/composer-normalize/pull/1056
 
 [@core23]: https://github.com/core23
 [@dependabot]: https://github.com/dependabot