+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`2.30.0...main`][2.30.0...main].
 
+### Fixed
+
+- Required `ergebnis/json-normalizer:^4.0.1` ([#1060]), by [@localheinz]
+
 ## [`2.30.0`][2.30.0]
 
 For a full diff see [`2.29.0...2.30.0`][2.29.0...2.30.0].
@@ -1008,6 +1012,7 @@ For a full diff see [`81bc3a8...0.1.0`][81bc3a8...0.1.0].
 [#1008]: https://github.com/ergebnis/composer-normalize/pull/1008
 [#1020]: https://github.com/ergebnis/composer-normalize/pull/1020
 [#1056]: https://github.com/ergebnis/composer-normalize/pull/1056
+[#1060]: https://github.com/ergebnis/composer-normalize/pull/1060
 
 [@core23]: https://github.com/core23
 [@dependabot]: https://github.com/dependabot